requirements.txt
requests
python-telegram-bot
pillow
python-dotenv